dbeaver/pro#8047 moved data provider
