+++ plugins/org.jkiss.dbeaver.model.sql/plugin.xml
@@ -52,19 +52,19 @@
         </generator>
         <!-- Result set (data) SQL generators-->
         <generator id="dataSelect" class="org.jkiss.dbeaver.model.sql.generator.resultset.SQLGeneratorSelectFromData" label="SELECT .. WHERE .. =" description="SELECT from table" order="1">
-            <objectType name="org.jkiss.dbeaver.model.DBDDataProvider"/>
+            <objectType name="org.jkiss.dbeaver.model.data.DBDDataProvider"/>
         </generator>
         <generator id="dataSelectMany" class="org.jkiss.dbeaver.model.sql.generator.resultset.SQLGeneratorSelectManyFromData" label="SELECT .. WHERE .. IN" description="SELECT from table IN" order="2">
-            <objectType name="org.jkiss.dbeaver.model.DBDDataProvider"/>
+            <objectType name="org.jkiss.dbeaver.model.data.DBDDataProvider"/>
         </generator>
         <generator id="dataInsert" class="org.jkiss.dbeaver.model.sql.generator.resultset.SQLGeneratorInsertFromData" label="INSERT" description="INSERT into table" order="3">
-            <objectType name="org.jkiss.dbeaver.model.DBDDataProvider"/>
+            <objectType name="org.jkiss.dbeaver.model.data.DBDDataProvider"/>
         </generator>
         <generator id="dataUpdate" class="org.jkiss.dbeaver.model.sql.generator.resultset.SQLGeneratorUpdateFromData" label="UPDATE" description="UPDATE table" order="4">
-            <objectType name="org.jkiss.dbeaver.model.DBDDataProvider"/>
+            <objectType name="org.jkiss.dbeaver.model.data.DBDDataProvider"/>
         </generator>
         <generator id="dataDeleteByUniqueKey" class="org.jkiss.dbeaver.model.sql.generator.resultset.SQLGeneratorDeleteFromData" label="DELETE by Unique Key" description="DELETE from table" order="5">
-            <objectType name="org.jkiss.dbeaver.model.DBDDataProvider"/>
+            <objectType name="org.jkiss.dbeaver.model.data.DBDDataProvider"/>
         </generator>
     </extension>
